Refactor tests.rs into multiple test modules
